+++ src/components/HomePage/modules/Banner.js
@@ -1,16 +1,9 @@
 import React, { useState, useEffect } from 'react';
 
 import { Stack, Text } from '@kiwicom/orbit-components/lib';
+import BannerImage from '../../ImageHolder/BannerImage';
 import SearchBar from './SearchBar';
 import styled from 'styled-components';
-
-const ImageContainer = styled.div`
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  object-fit: cover;
-  filter: brightness(70%);
-`;
 
 const TitleArea = styled.div`
   position: absolute;
@@ -35,26 +28,9 @@ const BannerContainer = styled.div`
   position: relative;
 `;
 
-const styles = {
-  bannerImage: {
-    width: '100%',
-    height: '100%',
-    position: 'absolute',
-    objectFit: 'cover',
-    filter: 'brightness(70%)',
-  },
-};
-
 const title = 'GiftForGood.sg';
 const subTitle = 'Giving back to the society that needs your help';
-
-const BannerImage = () => {
-  return (
-    <div>
-      <img style={styles.bannerImage} src="/assets/wishes-banner.jpg" />
-    </div>
-  );
-};
+const bannerImageUrl = '/assets/wishes-banner.jpg';
 
 const BannerText = () => {
   return (
@@ -69,7 +45,7 @@ const Banner = () => {
   return (
     <BannerContainer>
       <Stack align="center">
-        <BannerImage />
+        <BannerImage imageUrl={bannerImageUrl} />
         <BannerText />
         <SearchBar />
       </Stack>
