+++ src/components/HomePage/modules/Categories.js
@@ -1,20 +1,12 @@
 import React, { useState, useEffect } from 'react';
-import { Stack, Text } from '@kiwicom/orbit-components/lib';
-import Desktop from '@kiwicom/orbit-components/lib/Desktop';
-import Mobile from '@kiwicom/orbit-components/lib/Mobile';
+import { useRouter } from 'next/router';
+import { Stack } from '@kiwicom/orbit-components/lib';
 import api from '../../../../utils/api/index';
+import SquareImageBox from '../../ImageHolder/SquareImageBox';
 import styled from 'styled-components';
 
 const Container = styled.div`
   margin: 0 auto;
-`;
-
-const SquareBox = styled.div`
-  width: calc(75px + 2vw);
-  height: calc(75px + 2vw);
-  min-width: 75px;
-  min-height: 75px;
-  position: relative;
 `;
 
 const ResizableTitle = styled.div`
@@ -23,44 +15,11 @@ const ResizableTitle = styled.div`
   margin-bottom: 10px;
 `;
 
-const CaptionInBox = styled.div`
-  position: absolute;
-  top: 80%;
-  left: 10%;
-`;
-
 const ScrollableRow = styled.div`
   width: fit-content;
   max-width: 95vw;
   overflow-x: scroll;
 `;
-
-const styles = {
-  imageStyle: {
-    width: '100%',
-    height: '100%',
-    objectFit: 'cover',
-    filter: 'brightness(70%)',
-    borderRadius: '1vw',
-  },
-};
-
-const SquareImageBox = ({ imageUrl, caption, captionSize, captionType }) => {
-  return (
-    <SquareBox>
-      <img style={styles.imageStyle} src={imageUrl} />
-      <CaptionInBox>
-        {/* 
-          captionType: primary, secondary, info, success, warning, critical, white
-          captionSize: small, normal, large
-        */}
-        <Text type={captionType || 'white'} size={captionSize || 'small'}>
-          {caption}
-        </Text>
-      </CaptionInBox>
-    </SquareBox>
-  );
-};
 
 const dummyCategoriesData = [
   {
@@ -131,12 +90,22 @@ const Categories = () => {
     // setCategoriesObj(api.categories.getAll());
   });
 
-  const CategoryRow = () => {
+  const RowOfCategories = () => {
+    const router = useRouter();
     return (
       <Stack direction="row" align="center">
-        {dummyCategoriesData.map((category) => (
-          <SquareImageBox key={category.id} imageUrl={category.imageUrl} caption={category.name} captionSize="small" />
-        ))}
+        {dummyCategoriesData.map((category) => {
+          const href = '/category/' + category.name;
+          const handleClick = (event) => {
+            event.preventDefault();
+            router.push(href);
+          };
+          return (
+            <a href={href} onClick={handleClick} key={category.id}>
+              <SquareImageBox imageUrl={category.imageUrl} caption={category.name} />
+            </a>
+          );
+        })}
       </Stack>
     );
   };
@@ -145,7 +114,7 @@ const Categories = () => {
     <Container>
       <ResizableTitle>Explore GiftForGood</ResizableTitle>
       <ScrollableRow>
-        <CategoryRow />
+        <RowOfCategories />
       </ScrollableRow>
     </Container>
   );
